change item_id type on dummy schema

+++ spec/dummy_app/db/migrate/20110208155312_set_up_test_tables.rb
@@ -73,7 +73,7 @@ class SetUpTestTables < ::ActiveRecord::Migration::Current
 
     create_table :versions, versions_table_options do |t|
       t.string   :item_type, **item_type_options(null: false)
-      t.integer  :item_id, null: false
+      t.bigint   :item_id, null: false
       t.string   :item_subtype, **item_type_options(null: true)
       t.string   :event, null: false
       t.string   :whodunnit
